requirements.txt
# requirements.txt

# Data manipulation and analysis
pandas==1.5.3
numpy==1.23.5

# Machine Learning
scikit-learn==1.2.2
xgboost==1.7.5

# Data visualization
matplotlib==3.7.1
seaborn==0.12.2

# Pipeline and workflow management
prefect==2.10.8

# Model tracking and experiment management
mlflow==2.3.1


pyyaml==5.4.1
joblib==1.0.1